tinyfpga: update for modern litex
Move the default SRAM size for the target (ie, TinyFPGA BX) into
the Makefile, as litex now has a (non-zero, but smaller) default
SRAM size.  So our desired default is in the Makefile, passed in
via command line arguments to make.

Also capture the specified ROM size, but force it to be in the
spiflash instead of in the FPGA block RAM.

Remove toolchain overrides for nextpnr, are these are now the
default for litex (and the layout of the commands to run has
changed in litex anyway).